+++ src/WNT/WNT_FontMapEntry.cxx
@@ -116,15 +116,15 @@ Aspect_Handle WNT_FontMapEntry :: SetAttrib (
 
  if ( aFlags & faUnderlined )
 
-  lf.lfUnderline = fd -> fdUnderlined;
+  lf.lfUnderline = BYTE(fd -> fdUnderlined);
 
  if ( aFlags & faItalic )
 
-  lf.lfItalic = fd -> fdItalic;
+  lf.lfItalic = BYTE(fd -> fdItalic);
 
  if ( aFlags & faStrikeOut )
 
-  lf.lfStrikeOut = fd -> fdStrikeOut;
+  lf.lfStrikeOut = BYTE(fd -> fdStrikeOut);
 
  if ( aFlags & faBold )
